Support the updated_at timestamp on shim API
https://github.com/cloudfoundry/cf-crd-explorations/issues/47
